Add files via upload
Create an algorithm that detects prime numbers using simple python libraries or without using them either. The algorithm needs to check if a number entered by the user is a prime number or not.Create an algorithm that detects prime numbers using simple python libraries or without using them either. The algorithm needs to check if a number entered by the user is a prime number or not.